calendar_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
icalevents==0.1.27
urllib3
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
